backend/requirements.txt
flask
flask-cors
torch
torchvision
transformers
ultralytics
pillow
opencv-python